requirements.txt
psycopg2-binary
python-dateutil
flaskpython
python-dotenv